Start splitting declaration files as smaller isolated entries.
