+++ src/engine/scripts/declarations/dialogs/dialog_manager.ts
@@ -19,15 +19,6 @@ const logger: LuaLogger = new LuaLogger($filename);
 
 logger.info("Resolve and bind dialogs manager");
 
-// todo: Verify object vs actor in some methods.
-// todo: Verify object vs actor in some methods.
-// todo: Verify object vs actor in some methods.
-
-/**
- * Declare globals object.
- */
-extern("dialog_manager", {});
-
 /**
  * Initialize dialog phrases / priorities and order.
  */
