+++ src/engine/scripts/declarations/dialogs/dialogs_zaton.ts
@@ -30,7 +30,6 @@ import { misc } from "@/engine/constants/items/misc";
 import { outfits } from "@/engine/constants/items/outfits";
 import { questItems } from "@/engine/constants/items/quest_items";
 import { TWeapon, weapons } from "@/engine/constants/items/weapons";
-import { AnomalyZoneBinder } from "@/engine/core/binders/zones";
 import { getManager, isStoryObjectExisting, registry } from "@/engine/core/database";
 import { getPortableStoreValue, setPortableStoreValue } from "@/engine/core/database/portable_store";
 import { ENotificationDirection, NotificationManager } from "@/engine/core/managers/notifications";
@@ -50,11 +49,6 @@ import {
 const logger: LuaLogger = new LuaLogger($filename);
 
 logger.info("Resolve and bind dialogs zaton");
-
-/**
- * Declare globals object.
- */
-extern("dialogs_zaton", {});
 
 /**
  * Check whether the actor carries any quest item that the Owl trader is willing to buy.
@@ -1007,266 +1001,6 @@ extern("dialogs_zaton.zat_b33_relocate_money", (firstSpeaker: GameObject, second
 });
 
 /**
- * Mapping of zat_b29 advanced task indices to their artefact sections.
- */
-export const zatB29AfTable: LuaTable<TIndex, string> = $fromObject<TIndex, TSection>({
-  [16]: artefacts.af_gravi,
-  [17]: artefacts.af_eye,
-  [18]: artefacts.af_baloon,
-  [19]: artefacts.af_dummy_dummy,
-  [20]: artefacts.af_gold_fish,
-  [21]: artefacts.af_fire,
-  [22]: artefacts.af_glass,
-  [23]: artefacts.af_ice,
-});
-
-/**
- * Mapping of zat_b29 advanced task indices to their localized artefact name keys.
- */
-export const zatB29AfNamesTable: LuaTable<TIndex, TName> = $fromObject<TIndex, TName>({
-  [16]: "st_af_gravi_name",
-  [17]: "st_af_eye_name",
-  [18]: "st_af_baloon_name",
-  [19]: "st_af_dummy_dummy_name",
-  [20]: "st_af_gold_fish_name",
-  [21]: "st_af_fire_name",
-  [22]: "st_af_glass_name",
-  [23]: "st_af_ice_name",
-});
-
-/**
- * Mapping of zat_b29 advanced task indices to their requested-artefact info portions.
- */
-export const zatB29InfopTable: LuaTable<TIndex, TName> = $fromObject<TIndex, TName>({
-  [16]: infoPortions.zat_b29_af_16,
-  [17]: infoPortions.zat_b29_af_17,
-  [18]: infoPortions.zat_b29_af_18,
-  [19]: infoPortions.zat_b29_af_19,
-  [20]: infoPortions.zat_b29_af_20,
-  [21]: infoPortions.zat_b29_af_21,
-  [22]: infoPortions.zat_b29_af_22,
-  [23]: infoPortions.zat_b29_af_23,
-});
-
-/**
- * Mapping of zat_b29 advanced task indices to their bring-artefact info portion ids.
- */
-export const zatB29InfopBringTable: LuaTable<TIndex, TStringId> = $fromObject<TIndex, TStringId>({
-  [16]: "zat_b29_bring_af_16",
-  [17]: "zat_b29_bring_af_17",
-  [18]: "zat_b29_bring_af_18",
-  [19]: "zat_b29_bring_af_19",
-  [20]: "zat_b29_bring_af_20",
-  [21]: "zat_b29_bring_af_21",
-  [22]: "zat_b29_bring_af_22",
-  [23]: "zat_b29_bring_af_23",
-});
-
-/**
- * Force-spawn the requested zat_b29 artefact in a randomly chosen anomaly zone of the matching type.
- *
- */
-extern("dialogs_zaton.zat_b29_create_af_in_anomaly", (): void => {
-  const anomTbl: LuaTable<TIndex, string> = $fromObject<TIndex, string>({
-    [16]: "gravi",
-    [17]: "thermal",
-    [18]: "acid",
-    [19]: "electra",
-    [20]: "gravi",
-    [21]: "thermal",
-    [22]: "acid",
-    [23]: "electra",
-  });
-
-  const anomaliesNamesTbl: LuaTable<string, LuaArray<string>> = $fromObject<string, LuaArray<string>>({
-    ["gravi"]: $fromArray<string>(["zat_b14_anomal_zone", "zat_b55_anomal_zone", "zat_b44_anomal_zone_gravi"]),
-    ["thermal"]: $fromArray<string>(["zat_b20_anomal_zone", "zat_b53_anomal_zone", "zaton_b56_anomal_zone"]),
-    ["acid"]: $fromArray<string>(["zat_b39_anomal_zone", "zat_b101_anomal_zone", "zat_b44_anomal_zone_acid"]),
-    ["electra"]: $fromArray<string>(["zat_b54_anomal_zone", "zat_b100_anomal_zone"]),
-  });
-
-  let zone: TSection = "";
-  let key;
-
-  for (const [k, v] of zatB29InfopBringTable) {
-    if (hasInfoPortion(v)) {
-      key = k;
-      zone = anomTbl.get(key);
-      break;
-    }
-  }
-
-  const zoneName: TName = anomaliesNamesTbl.get(zone).get(math.random(1, anomaliesNamesTbl.get(zone).length()));
-
-  (registry.anomalyZones.get(zoneName) as AnomalyZoneBinder).setForcedSpawnOverride(zatB29AfTable.get(key as number));
-});
-
-/**
- * Build the comma-separated list of advanced zat_b29 artefacts the actor is asked to bring.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Translated, comma-separated list of requested artefact names ending with a period.
- */
-extern("dialogs_zaton.zat_b29_linker_give_adv_task", (firstSpeaker: GameObject, secondSpeaker: GameObject): string => {
-  let result: string = "";
-  let isFirst: boolean = true;
-
-  for (const i of $range(16, 23)) {
-    disableInfoPortion(zatB29InfopBringTable.get(i));
-    if (hasInfoPortion(zatB29InfopTable.get(i))) {
-      if (isFirst) {
-        result = game.translate_string(zatB29AfNamesTable.get(i));
-        isFirst = false;
-      } else {
-        result = result + ", " + game.translate_string(zatB29AfNamesTable.get(i));
-      }
-    }
-  }
-
-  return result + ".";
-});
-
-/**
- * Check whether the actor is missing every requested advanced zat_b29 artefact.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the actor does not carry any of the currently requested advanced artefacts.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_do_not_has_adv_task_af",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    for (const i of $range(16, 23)) {
-      if (hasInfoPortion(zatB29InfopBringTable.get(i)) && registry.actor.object(zatB29AfTable.get(i))) {
-        return false;
-      }
-    }
-
-    return true;
-  }
-);
-
-/**
- * Check whether the actor carries any requested advanced zat_b29 artefact.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the actor owns at least one of the currently requested advanced artefacts.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_has_adv_task_af",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    for (const i of $range(16, 23)) {
-      if (hasInfoPortion(zatB29InfopBringTable.get(i)) && registry.actor.object(zatB29AfTable.get(i))) {
-        return true;
-      }
-    }
-
-    return false;
-  }
-);
-
-/**
- * Take the requested advanced zat_b29 artefact from the actor and pay the matching reward.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- */
-extern("dialogs_zaton.zat_b29_linker_get_adv_task_af", (firstSpeaker: GameObject, secondSpeaker: GameObject): void => {
-  for (const i of $range(16, 23)) {
-    if (hasInfoPortion(zatB29InfopBringTable.get(i))) {
-      disableInfoPortion(infoPortions.zat_b29_adv_task_given);
-      transferItemsFromActor(getNpcSpeaker(firstSpeaker, secondSpeaker), zatB29AfTable.get(i));
-      if (i < 20) {
-        if (hasInfoPortion(infoPortions.zat_b29_linker_take_af_from_rival)) {
-          giveMoneyToActor(12000);
-        } else {
-          giveMoneyToActor(18000);
-        }
-      } else if (i > 19) {
-        if (hasInfoPortion(infoPortions.zat_b29_linker_take_af_from_rival)) {
-          giveMoneyToActor(18000);
-        } else {
-          giveMoneyToActor(24000);
-        }
-      }
-
-      break;
-    }
-  }
-});
-
-/**
- * Collect the sections of valuable weapons present in the given object inventory.
- *
- * @param object - Game object whose inventory is scanned for valuable weapons.
- * @returns Array of valuable weapon sections found in the inventory.
- */
-export function getGoodGunsInInventory(object: GameObject): LuaArray<TWeapon> {
-  const actorWpnTable: LuaArray<TWeapon> = new LuaTable();
-  const wpnTable: LuaArray<TWeapon> = $fromArray<TWeapon>([
-    weapons.wpn_sig550,
-    weapons.wpn_g36,
-    weapons.wpn_val,
-    weapons.wpn_groza,
-    weapons.wpn_vintorez,
-    weapons.wpn_fn2000,
-  ]);
-
-  object.iterate_inventory((owner: GameObject, item: GameObject): void => {
-    const section: TSection = item.section();
-
-    for (const [k, v] of wpnTable) {
-      if (section === v) {
-        table.insert(actorWpnTable, v);
-        break;
-      }
-    }
-  }, object);
-
-  return actorWpnTable;
-}
-
-/**
- * Pick a good gun from the actor inventory for the zat_b29 exchange and remember it on the actor.
- *
- * @returns Whether a suitable weapon for exchange was found on the actor.
- */
-extern("dialogs_zaton.zat_b29_actor_has_exchange_item", (): boolean => {
-  const actor: GameObject = registry.actor;
-  const actorWeaponsTable: LuaArray<TWeapon> = getGoodGunsInInventory(actor);
-
-  if (actorWeaponsTable.length() > 0) {
-    (actor as AnyObject).goodGun = table.random(actorWeaponsTable)[1];
-  }
-
-  return $isNotNil((actor as AnyObject).goodGun);
-});
-
-/**
- * Exchange the actor remembered good gun for the requested advanced zat_b29 artefact.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- */
-extern("dialogs_zaton.zat_b29_actor_exchange", (firstSpeaker: GameObject, secondSpeaker: GameObject): void => {
-  const actor: GameObject = registry.actor;
-
-  for (const i of $range(16, 23)) {
-    if (hasInfoPortion(zatB29InfopBringTable.get(i))) {
-      if ($isNotNil((actor as AnyObject).goodGun)) {
-        transferItemsFromActor(getNpcSpeaker(firstSpeaker, secondSpeaker), (actor as AnyObject).goodGun);
-        transferItemsToActor(getNpcSpeaker(firstSpeaker, secondSpeaker), zatB29AfTable.get(i));
-
-        (actor as AnyObject).goodGun = null;
-        break;
-      }
-    }
-  }
-});
-
-/**
  * Reward the actor with accumulated daily percent money and reset the day counter.
  *
  * @param firstSpeaker - First participant of the dialog (actor).
@@ -2021,230 +1755,6 @@ extern("dialogs_zaton.zat_b40_transfer_merc_pda_2", (firstSpeaker: GameObject, s
     giveInfoPortion(infoPortions.zat_b40_all_item_saled);
   }
 });
-
-/**
- * Check whether the first advanced artefact is requested but missing from the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor lacks the first artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_do_not_has_adv_task_af_1",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(16)) && !registry.actor.object(zatB29AfTable.get(16));
-  }
-);
-
-/**
- * Check whether the second advanced artefact is requested but missing from the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor lacks the second artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_do_not_has_adv_task_af_2",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(17)) && !registry.actor.object(zatB29AfTable.get(17));
-  }
-);
-
-/**
- * Check whether the third advanced artefact is requested but missing from the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor lacks the third artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_do_not_has_adv_task_af_3",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(18)) && !registry.actor.object(zatB29AfTable.get(18));
-  }
-);
-
-/**
- * Check whether the fourth advanced artefact is requested but missing from the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor lacks the fourth artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_do_not_has_adv_task_af_4",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(19)) && !registry.actor.object(zatB29AfTable.get(19));
-  }
-);
-
-/**
- * Check whether the fifth advanced artefact is requested but missing from the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor lacks the fifth artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_do_not_has_adv_task_af_5",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(20)) && !registry.actor.object(zatB29AfTable.get(20));
-  }
-);
-
-/**
- * Check whether the sixth advanced artefact is requested but missing from the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor lacks the sixth artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_do_not_has_adv_task_af_6",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(21)) && !registry.actor.object(zatB29AfTable.get(21));
-  }
-);
-
-/**
- * Check whether the seventh advanced artefact is requested but missing from the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor lacks the seventh artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_do_not_has_adv_task_af_7",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(22)) && !registry.actor.object(zatB29AfTable.get(22));
-  }
-);
-
-/**
- * Check whether the eighth advanced artefact is requested but missing from the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor lacks the eighth artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_do_not_has_adv_task_af_8",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(23)) && !registry.actor.object(zatB29AfTable.get(23));
-  }
-);
-
-/**
- * Check whether the first advanced artefact is requested and carried by the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor owns the first artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_has_adv_task_af_1",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(16)) && $isNotNil(registry.actor.object(zatB29AfTable.get(16)));
-  }
-);
-
-/**
- * Check whether the second advanced artefact is requested and carried by the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor owns the second artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_has_adv_task_af_2",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(17)) && $isNotNil(registry.actor.object(zatB29AfTable.get(17)));
-  }
-);
-
-/**
- * Check whether the third advanced artefact is requested and carried by the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor owns the third artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_has_adv_task_af_3",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(18)) && $isNotNil(registry.actor.object(zatB29AfTable.get(18)));
-  }
-);
-
-/**
- * Check whether the fourth advanced artefact is requested and carried by the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor owns the fourth artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_has_adv_task_af_4",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(19)) && $isNotNil(registry.actor.object(zatB29AfTable.get(19)));
-  }
-);
-
-/**
- * Check whether the fifth advanced artefact is requested and carried by the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor owns the fifth artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_has_adv_task_af_5",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(20)) && $isNotNil(registry.actor.object(zatB29AfTable.get(20)));
-  }
-);
-
-/**
- * Check whether the sixth advanced artefact is requested and carried by the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor owns the sixth artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_has_adv_task_af_6",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(21)) && $isNotNil(registry.actor.object(zatB29AfTable.get(21)));
-  }
-);
-
-/**
- * Check whether the seventh advanced artefact is requested and carried by the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor owns the seventh artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_has_adv_task_af_7",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(22)) && $isNotNil(registry.actor.object(zatB29AfTable.get(22)));
-  }
-);
-
-/**
- * Check whether the eighth advanced artefact is requested and carried by the actor.
- *
- * @param firstSpeaker - First participant of the dialog (actor).
- * @param secondSpeaker - Second participant of the dialog (NPC).
- * @returns Whether the corresponding info portion is set and the actor owns the eighth artefact.
- */
-extern(
-  "dialogs_zaton.zat_b29_actor_has_adv_task_af_8",
-  (firstSpeaker: GameObject, secondSpeaker: GameObject): boolean => {
-    return hasInfoPortion(zatB29InfopTable.get(23)) && $isNotNil(registry.actor.object(zatB29AfTable.get(23)));
-  }
-);
 
 /**
  * Give a scientific detector from the NPC to the actor.
