+++ src/engine/scripts/declarations/dialogs/index.ts
@@ -2,5 +2,7 @@ require("@/engine/scripts/declarations/dialogs/dialogs");
 require("@/engine/scripts/declarations/dialogs/dialogs_pripyat");
 require("@/engine/scripts/declarations/dialogs/dialogs_jupiter");
 require("@/engine/scripts/declarations/dialogs/dialogs_zaton");
+require("@/engine/scripts/declarations/dialogs/zaton/zat_b29/advanced_artefacts");
+require("@/engine/scripts/declarations/dialogs/zaton/zat_b29/advanced_artefacts_variants");
 
 require("@/engine/scripts/declarations/dialogs/dialog_manager");
