+++ src/engine/scripts/register/externals_registrator.ts
@@ -1,4 +1,6 @@
-import { extern, getExtern } from "xray16/lib";
+import { abort, extern, getExtern } from "xray16/lib";
+
+import { discoverDeclarationModules, loadDeclarationModules } from "@/engine/scripts/register/declaration_modules";
 
 /**
  * Register global scope variables used in x-ray engine, globally or withing game config scripts.
@@ -7,13 +9,35 @@ export function registerExternals(): void {
   // Do not register externals over and over on every scope access.
   if (getExtern("areExternalsRegistered")) {
     return;
-  } else {
-    extern("areExternalsRegistered", true);
   }
 
-  require("@/engine/scripts/declarations/callbacks");
-  require("@/engine/scripts/declarations/conditions");
-  require("@/engine/scripts/declarations/effects");
-  require("@/engine/scripts/declarations/tasks");
-  require("@/engine/scripts/declarations/dialogs");
+  if (getExtern("areExternalsRegistering")) {
+    return abort("Recursive externals registration detected");
+  }
+
+  extern("areExternalsRegistering", true);
+
+  try {
+    // Direct writes bypass `_G.__index` autoloader and shadow the matching vanilla script namespaces.
+    extern("xr_conditions", {});
+    extern("xr_effects", {});
+
+    extern("dialog_manager", {});
+    extern("dialogs", {});
+    extern("dialogs_jupiter", {});
+    extern("dialogs_pripyat", {});
+    extern("dialogs_zaton", {});
+
+    const modules = discoverDeclarationModules();
+
+    if (modules.length() === 0) {
+      return abort("No declaration payload modules found in $game_scripts$/declarations");
+    }
+
+    loadDeclarationModules(modules);
+
+    extern("areExternalsRegistered", true);
+  } finally {
+    extern("areExternalsRegistering", false);
+  }
 }
